chore: refactor images file structure
