+++ src/features/components/templates/navigation/main-nav/content/platform/index.ts
@@ -3,12 +3,12 @@ import DerivcTrader from 'images/svg/components/component_deriv_ctrader_icon.svg
 import BinaryBot from 'images/svg/binarybot_icon.svg'
 import DerivX from 'images/svg/deriv_x_icon.svg'
 import CFD from 'images/svg/components/component_cfds_icon.svg'
-import Multipliers from 'images/svg/custom/rebranding/multipliers-icon.svg'
-import Options from 'images/svg/custom/rebranding/options-icon.svg'
-import Smarttrader from 'images/svg/custom/rebranding/smarttrader-icon.svg'
+import Multipliers from 'images/svg/components/component_multipliers_icon.svg'
+import Options from 'images/svg/components/component_options_icon.svg'
+import Smarttrader from 'images/svg/smart_trader_icon.svg'
 import DerivGo from 'images/svg/deriv_go_icon.svg'
-import DMT5 from 'images/svg/custom/rebranding/dmt5-icon.svg'
-import DTrader from 'images/svg/custom/rebranding/dtrader-icon.svg'
+import DMT5 from 'images/svg/dmt5_icon.svg'
+import DTrader from 'images/svg/dtrader_icon.svg'
 import DerivEZ from 'images/svg/deriv_ez_icon.svg'
 import DerivBot from 'images/svg/deriv_bot_icon.svg'
