+++ src/features/pages/home/our-platforms/data.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { StaticImage } from 'gatsby-plugin-image'
 import { SmartPlatformItem } from './types'
-import DTraderIcon from 'images/svg/custom/rebranding/dtrader-icon.svg'
-import DMT5Icon from 'images/svg/custom/rebranding/dmt5-icon.svg'
+import DTraderIcon from 'images/svg/dtrader_icon.svg'
+import DMT5Icon from 'images/svg/dmt5_icon.svg'
 import DerivGOIcon from 'images/svg/deriv_go_icon.svg'
 import DerivXIcon from 'images/svg/deriv_x_icon.svg'
 import DBotIcon from 'images/svg/deriv_bot_icon.svg'
 import BinaryBotIcon from 'images/svg/binarybot-icon.svg'
-import SmartTraderIcon from 'images/svg/custom/rebranding/smarttrader-icon.svg'
+import SmartTraderIcon from 'images/svg/smart_trader_icon.svg'
 import DerivEzIcon from 'images/svg/deriv_ez_icon.svg'
 import Link from 'features/components/atoms/link'
 import {
