+++ src/images/svg/deriv-ez/index.ts
@@ -3,7 +3,7 @@ export { default as DerivEzLogoRed } from './deriv-ez-logo-red.svg'
 export { default as DerivEzLogo } from './deriv-ez-logo.svg'
 export { default as DerivEzQR } from './derivez-qr.svg'
 export { default as Dmt5Icon } from './dmt5-icon.svg'
-export { default as DtIcon } from '../custom/rebranding/dtrader-icon.svg'
+export { default as DtIcon } from '../dtrader_icon.svg'
 export { default as DxIcon } from '../deriv_x_icon.svg'
 export { default as EasyToStart } from './easy-to-start.svg'
 export { default as FavouriteAssets } from './favourite-assets.svg'
